Response file input and correction file inputs refactored
