+++ src/app/page.tsx
@@ -5,6 +5,8 @@ import { open } from "@tauri-apps/api/dialog";
 import { invoke, convertFileSrc } from "@tauri-apps/api/tauri";
 import CroppingResizingViewSettings from "./cropping-resizing-view-settings";
 import Navigation from "./navigation";
+import Response_and_correction from "./response_and_correction";
+import { Paths } from "./string_functions";
 
 const DEBUG = true;
 const fakePipeline = false;
@@ -20,15 +22,12 @@ export default function Home() {
         if (DEBUG) {
           console.log("OS platform successfully queried:", platform);
         }
-
         // Default path for macOS and Linux
         let defaultPath = "/usr/local/bin";
-
         // If platform is windows, update default path
         if (osPlatform === "windows") {
           defaultPath = "C:\\bin";
         }
-
         // Update settings
         setSettings({
           radiancePath: defaultPath,
@@ -60,13 +59,6 @@ export default function Home() {
   const [progressButton, setProgressButton] = useState<boolean>(false);
   const [processError, setProcessError] = useState<boolean>(false);
 
-  // Error checking display
-  const [response_error, set_response_error] = useState<boolean>(false);
-  const [fe_error, set_fe_error] = useState<boolean>(false);
-  const [v_error, set_v_error] = useState<boolean>(false);
-  const [nd_error, set_nd_error] = useState<boolean>(false);
-  const [cf_error, set_cf_error] = useState<boolean>(false);
-
   // PATH AND FILE INFORMATION
 
   // const [imgDirs, setImgDirs] = useState<any[]>([]);
@@ -81,16 +73,11 @@ export default function Home() {
   let selected: any | any[] = [];
   // Holds the temporary asset paths selected by the user during the dialog function
   let assets: any[] = [];
-  let response: any = "";
   const [responsePaths, setResponsePaths] = useState<string>("");
   // Correction files fe = fish eye, v= vignetting, nd = neutral density, cf = calibration factor
-  let fe_correction: any = "";
   const [fe_correctionPaths, set_fe_correctionPaths] = useState<string>("");
-  let v_correction: any = "";
   const [v_correctionPaths, set_v_correctionPaths] = useState<string>("");
-  let nd_correction: any = "";
   const [nd_correctionPaths, set_nd_correctionPaths] = useState<string>("");
-  let cf_correction: any = "";
   const [cf_correctionPaths, set_cf_correctionPaths] = useState<string>("");
   const [images, setImages] = useState<File[]>([]);
 
@@ -108,28 +95,6 @@ export default function Home() {
     updatedSettings[event.currentTarget.name] = event.currentTarget.value;
     setSettings(updatedSettings);
   };
-
-  // FILE STRING FUNCTIONS
-
-  function Paths(path: string) {
-    for (let i = 0; i < path.length; i++) {
-      if (path[i] == "/" || path[i] == "\\") {
-        path = path.slice(i + 1);
-        i = -1;
-      }
-    }
-    return path;
-  }
-
-  function Extensions(ext: string) {
-    for (let i = 0; i < ext.length; i++) {
-      if (ext[i] == ".") {
-        ext = ext.slice(i + 1);
-        i = -1;
-      }
-    }
-    return ext;
-  }
 
   // Reset progress
   function ResetProgress() {
@@ -179,103 +144,6 @@ export default function Home() {
     }
   }
 
-  async function dialogResponse() {
-    response = await open({
-      multiple: true,
-    });
-    if (response === null) {
-      // user cancelled the selection
-    } else {
-      console.log("Extension " + Extensions(response[0]));
-      set_response_error(false);
-      if (Extensions(response[0]) !== "rsp") {
-        set_response_error(true);
-      } else {
-        setResponsePaths(response[0]);
-      }
-    }
-    if (DEBUG) {
-      console.log("response: ", response);
-    }
-  }
-
-  async function dialogFE() {
-    fe_correction = await open({
-      multiple: true,
-    });
-    if (fe_correction === null) {
-      // user cancelled the selection
-    } else {
-      console.log(fe_correction[0]);
-      set_fe_error(false);
-      if (Extensions(fe_correction[0]) !== "cal") {
-        set_fe_error(true);
-      } else {
-        set_fe_correctionPaths(fe_correction[0]);
-      }
-    }
-    if (DEBUG) {
-      console.log("fe_correction: ", fe_correction);
-    }
-  }
-
-  async function dialogV() {
-    v_correction = await open({
-      multiple: true,
-    });
-    if (v_correction === null) {
-      // user cancelled the selection
-    } else {
-      set_v_error(false);
-      if (Extensions(v_correction[0]) !== "cal") {
-        set_v_error(true);
-      } else {
-        set_v_correctionPaths(v_correction[0]);
-      }
-    }
-    if (DEBUG) {
-      console.log("v_correction: ", v_correction);
-    }
-  }
-
-  async function dialogND() {
-    nd_correction = await open({
-      multiple: true,
-    });
-    if (nd_correction === null) {
-      // user cancelled the selection
-    } else {
-      set_nd_error(false);
-      if (Extensions(nd_correction[0]) !== "cal") {
-        set_nd_error(true);
-      } else {
-        set_nd_correctionPaths(nd_correction[0]);
-      }
-    }
-    if (DEBUG) {
-      console.log("nd_correction: ", nd_correction);
-    }
-  }
-
-  async function dialogCF() {
-    cf_correction = await open({
-      multiple: true,
-    });
-    if (cf_correction === null) {
-      // user cancelled the selection
-    } else {
-      set_cf_error(false);
-      if (Extensions(cf_correction[0]) !== "cal") {
-        set_cf_error(true);
-      } else {
-        set_cf_correctionPaths(cf_correction[0]);
-      }
-    }
-    if (DEBUG) {
-      console.log("cf_correction: ", cf_correction);
-    }
-  }
-
   // DELETE FUNCTIONS
 
   const reset = () => {
@@ -299,26 +167,6 @@ export default function Home() {
     setImages(updatedImages);
     setDevicePaths(updatedDevicePaths);
     setAssetPaths(updatedAssetPaths);
-  };
-
-  const handleResponseDelete = () => {
-    setResponsePaths("");
-  };
-
-  const handle_fe_delete = () => {
-    set_fe_correctionPaths("");
-  };
-
-  const handle_v_delete = () => {
-    set_v_correctionPaths("");
-  };
-
-  const handle_nd_delete = () => {
-    set_nd_correctionPaths("");
-  };
-
-  const handle_cf_delete = () => {
-    set_cf_correctionPaths("");
   };
 
   if (DEBUG) {
@@ -409,24 +257,24 @@ export default function Home() {
     <main className="bg-white flex min-h-screen flex-col items-center justify-between text-black">
       <div>
         <Navigation
-        responsePaths = {responsePaths}
-        fe_correctionPaths = {fe_correctionPaths}
-        v_correctionPaths = {v_correctionPaths}
-        nd_correctionPaths = {nd_correctionPaths}
-        cf_correctionPaths = {cf_correctionPaths}
-        viewSettings = {viewSettings}
-        setConfig = {setConfig}
-        settings = {settings}
-        setSettings = {setSettings}
-        handleSettingsChange = {handleSettingsChange}
-        handleGenerateHDRImage = {handleGenerateHDRImage}
-        showProgress = {showProgress}
-        fakePipeline = {fakePipeline}
-        setProgressButton = {setProgressButton}
-        setProcessError = {setProcessError}
-        progressButton = {progressButton}
-        processError = {processError}
-        ResetProgress = {ResetProgress}/>
+          responsePaths={responsePaths}
+          fe_correctionPaths={fe_correctionPaths}
+          v_correctionPaths={v_correctionPaths}
+          nd_correctionPaths={nd_correctionPaths}
+          cf_correctionPaths={cf_correctionPaths}
+          viewSettings={viewSettings}
+          setConfig={setConfig}
+          settings={settings}
+          setSettings={setSettings}
+          handleSettingsChange={handleSettingsChange}
+          handleGenerateHDRImage={handleGenerateHDRImage}
+          showProgress={showProgress}
+          fakePipeline={fakePipeline}
+          setProgressButton={setProgressButton}
+          setProcessError={setProcessError}
+          progressButton={progressButton}
+          processError={processError}
+          ResetProgress={ResetProgress} />
         <div className="w-3/4 ml-auto pl-3">
           <h1 className="font-bold pt-10">Configuration</h1>
           <button
@@ -494,122 +342,24 @@ export default function Home() {
               </div>
             </>
           )}
-          <h2 className="font-bold pt-5" id="response">
-            Response File
-          </h2>
-          <button
-            onClick={dialogResponse}
-            className="bg-gray-300 hover:bg-gray-400 text-gray-700 font-semibold py-1 px-2 border-gray-400 rounded"
-          >
-            Select File
-          </button>
-          <div>
-            {response_error && (
-              <div>
-                <p>Please only enter files ending in rsp</p>
-              </div>
-            )}
-            {responsePaths && (
-              <div>
-                {Paths(responsePaths)}{" "}
-                <button onClick={() => handleResponseDelete()}>Delete</button>
-              </div>
-            )}
-          </div>
+
           <div id="c_r_v">
             <CroppingResizingViewSettings
               viewSettings={viewSettings}
               handleChange={handleViewSettingsChange}
             />
-          </div>
-          <h2 className="font-bold pt-5" id="fe">
-            Fish Eye Correction
-          </h2>
-          <button
-            onClick={dialogFE}
-            className="bg-gray-300 hover:bg-gray-400 text-gray-700 font-semibold py-1 px-2 border-gray-400 rounded"
-          >
-            Select File
-          </button>
-          <div>
-            {fe_error && (
-              <div>
-                <p>Please only enter files ending in cal</p>
-              </div>
-            )}
-            {fe_correctionPaths && (
-              <div>
-                {Paths(fe_correctionPaths)}{" "}
-                <button onClick={() => handle_fe_delete()}>Delete</button>
-              </div>
-            )}
-          </div>
-          <h2 className="font-bold pt-5" id="v">
-            Vignetting Correction
-          </h2>
-          <button
-            onClick={dialogV}
-            className="bg-gray-300 hover:bg-gray-400 text-gray-700 font-semibold py-1 px-2 border-gray-400 rounded"
-          >
-            Select File
-          </button>
-          <div>
-            {v_error && (
-              <div>
-                <p>Please only enter files ending in cal</p>
-              </div>
-            )}
-            {v_correctionPaths && (
-              <div>
-                {Paths(v_correctionPaths)}{" "}
-                <button onClick={() => handle_v_delete()}>Delete</button>
-              </div>
-            )}
-          </div>
-          <h2 className="font-bold pt-5" id="nd">
-            Neutral Density Correction
-          </h2>
-          <button
-            onClick={dialogND}
-            className="bg-gray-300 hover:bg-gray-400 text-gray-700 font-semibold py-1 px-2 border-gray-400 rounded"
-          >
-            Select File
-          </button>
-          <div>
-            {nd_error && (
-              <div>
-                <p>Please only enter files ending in cal</p>
-              </div>
-            )}
-            {nd_correctionPaths && (
-              <div>
-                {Paths(nd_correctionPaths)}{" "}
-                <button onClick={() => handle_nd_delete()}>Delete</button>
-              </div>
-            )}
-          </div>
-          <h2 className="font-bold pt-5" id="cf">
-            Calibration Factor Correction
-          </h2>
-          <button
-            onClick={dialogCF}
-            className="bg-gray-300 hover:bg-gray-400 text-gray-700 font-semibold py-1 px-2 border-gray-400 rounded"
-          >
-            Select File
-          </button>
-          <div>
-            {cf_error && (
-              <div>
-                <p>Please only enter files ending in cal</p>
-              </div>
-            )}
-            {cf_correctionPaths && (
-              <div>
-                {Paths(cf_correctionPaths)}{" "}
-                <button onClick={() => handle_cf_delete()}>Delete</button>
-              </div>
-            )}
-            <div className="pt-5"></div>
+            <Response_and_correction
+                  responsePaths = {responsePaths}
+                  fe_correctionPaths = {fe_correctionPaths}
+                  v_correctionPaths = {v_correctionPaths}
+                  nd_correctionPaths = {nd_correctionPaths}
+                  cf_correctionPaths = {cf_correctionPaths}
+                  setResponsePaths = {setResponsePaths}
+                  set_fe_correctionPaths = {set_fe_correctionPaths}
+                  set_v_correctionPaths = {set_v_correctionPaths}
+                  set_nd_correctionPaths = {set_nd_correctionPaths}
+                  set_cf_correctionPaths = {set_cf_correctionPaths}
+            />
           </div>
         </div>
       </div>
